Handle non-QP cases
Since, presumably, all model files will pass through the parser, we need to be able to detect when a model *doesn't* contain a quadratic element so that the rest of the parsing functions don't fire.

Instead, we just want to check for a QP element, and if none is found, pass it on to `Rglpk` and return the model object.

To that end, several private methods have been added:
- `check_quadratic` is a simple parser for detecting the QP pattern
- `isQP` crawls the character string until the objective portion and then calls `check_quadratic`, returning a boolean
- `processQP` now checks for a quadratic element immediately after reading the file path and will short-circuit the remainder of the parsers if no QP element is found
- finally, `%O>%` is a utility operator for evaluating the result of a combinator in a pipeline. Notably, it does not return a new combinator, but instead simply evaluates the RHS