Clear up naming issues with a big refactor
Since a lot of the functions have changed in scope and intended use, their names didn't always sync up with what they did or what they returned, so this is an attempt to clear that up.

The main entry point of the package is now `readQP`, which reads files, parses, creates model objects, etc.

The old `readQP`, which was just a utility function for reading files in as strings, is now `readModelFile` and is not exported.

`processQP` now has less responsibility. It receives a model string and assumes it has already been verified as a QP problem (`readQP` handles the verification now). In addition, `processQP` now returns a composite object containing the LP model AND the coefficient -> variable map.