Handle variations in the constraint expression
The CPLEX spec provides multiple options for expressing the constraint section of a model file: "The constraints section is introduced by the keyword subject to. This expression can also appear as such that, st, S.T., or ST. in any mix of upper- and lower-case characters. One of these expressions must precede the first constraint and be separated from it by at least one space."

Since the parser has a hard pivot on the beginning of the constraint section, all of these variations need to be accounted for in the parsers as well (they were originally hard coded to "Subject To").

Added a standalone pattern that captures all variations and refactored the parses to incorporate the pattern.